move composable to composables package and add tests
